move preparing of error data to separate module
