+++ src/services/index.js
@@ -6,6 +6,7 @@ const corsService = require('./cors.service')
 const helpers = require('./helpers')
 const grenacheClientService = require('./grenacheClient.service')
 const initWebSocketService = require('./initWebSocket.service')
+const prepareErrorDataService = require('./prepareErrorData.service')
 
 module.exports = {
   logService,
@@ -13,6 +14,7 @@ module.exports = {
   corsService,
   helpers,
   grenacheClientService,
-  initWebSocketService
+  initWebSocketService,
+  prepareErrorDataService
 
 }
